factorize internal error macro
