+++ meilisearch-http/src/index_controller/dump_actor/error.rs
@@ -1,8 +1,7 @@
 use meilisearch_error::{Code, ErrorCode};
 
-use crate::index_controller::{
-    update_actor::error::UpdateActorError, uuid_resolver::UuidResolverError,
-};
+use crate::index_controller::update_actor::error::UpdateActorError;
+use crate::index_controller::uuid_resolver::error::UuidResolverError;
 
 pub type Result<T> = std::result::Result<T, DumpActorError>;
