+++ meilisearch-http/src/index_controller/uuid_resolver/mod.rs
@@ -1,4 +1,5 @@
 mod actor;
+pub mod error;
 mod handle_impl;
 mod message;
 pub mod store;
@@ -6,14 +7,12 @@ pub mod store;
 use std::collections::HashSet;
 use std::path::PathBuf;
 
-use meilisearch_error::Code;
-use meilisearch_error::ErrorCode;
-use thiserror::Error;
 use uuid::Uuid;
 
 use actor::UuidResolverActor;
 use message::UuidResolveMsg;
 use store::UuidStore;
+use error::Result;
 
 #[cfg(test)]
 use mockall::automock;
@@ -23,7 +22,6 @@ pub use store::HeedUuidStore;
 
 const UUID_STORE_SIZE: usize = 1_073_741_824; //1GiB
 
-pub type Result<T> = std::result::Result<T, UuidResolverError>;
 
 #[async_trait::async_trait]
 #[cfg_attr(test, automock)]
@@ -35,47 +33,4 @@ pub trait UuidResolverHandle {
     async fn snapshot(&self, path: PathBuf) -> Result<HashSet<Uuid>>;
     async fn get_size(&self) -> Result<u64>;
     async fn dump(&self, path: PathBuf) -> Result<HashSet<Uuid>>;
-}
-
-#[derive(Debug, Error)]
-pub enum UuidResolverError {
-    #[error("Name already exist.")]
-    NameAlreadyExist,
-    #[error("Index \"{0}\" doesn't exist.")]
-    UnexistingIndex(String),
-    #[error("Badly formatted index uid: {0}")]
-    BadlyFormatted(String),
-    #[error("Internal error resolving index uid: {0}")]
-    Internal(Box<dyn std::error::Error + Sync + Send + 'static>),
-}
-
-macro_rules! internal_error {
-    ($($other:path), *) => {
-        $(
-            impl From<$other> for UuidResolverError {
-                fn from(other: $other) -> Self {
-                    Self::Internal(Box::new(other))
-                }
-            }
-        )*
-    }
-}
-
-internal_error!(
-    heed::Error,
-    uuid::Error,
-    std::io::Error,
-    tokio::task::JoinError,
-    serde_json::Error
-);
-
-impl ErrorCode for UuidResolverError {
-    fn error_code(&self) -> Code {
-        match self {
-            UuidResolverError::NameAlreadyExist => Code::IndexAlreadyExists,
-            UuidResolverError::UnexistingIndex(_) => Code::IndexNotFound,
-            UuidResolverError::BadlyFormatted(_) => Code::InvalidIndexUid,
-            UuidResolverError::Internal(_) => Code::Internal,
-        }
-    }
 }
