+++ meilisearch-http/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod data;
+#[macro_use]
 pub mod error;
 pub mod helpers;
 mod index;
